requirements.txt
absl-py==1.4.0
aiohttp==3.8.6
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
branca==0.6.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
cvzone==1.6.1
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
dict==2020.12.3
enchant==0.0.1
entrypoints==0.4
Flask==2.2.5
flatbuffers==2.0.7
folium==0.14.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gast==0.4.0
geocoder==1.38.1
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.57.0
gTTS==2.4.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.7.0
imutils==0.5.4
ipykernel==6.16.2
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==7.4.9
jupyter_core==4.12.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
lightning-utilities==0.10.0
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mediapipe==0.9.0.1
multidict==6.0.4
nest-asyncio==1.5.7
numpy==1.21.6
oauthlib==3.2.2
opencage==2.3.0
opencv-contrib-python==4.8.1.78
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==23.1
pandas==1.3.5
parso==0.8.3
phonenumbers==8.13.22
pickleshare==0.7.5
Pillow==9.5.0
playsound==1.3.0
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydotplus==2.0.2
pyenchant==3.2.2
pygame==2.5.2
Pygments==2.16.1
pyparsing==3.1.1
pypiwin32==223
python-dateutil==2.8.2
pytorch-lightning==1.9.5
pyttsx3==2.90
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
ratelim==0.1.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0.post7
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tf==1.0.0
threadpoolctl==3.1.0
tk==0.1.0
torch==1.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==1.26.16
values==2020.12.3
vboxapi==1.0
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.9.4
zipp==3.15.0
